Fix incorrect L classification for specific shifts
If the shift value was equal to or greater than the width of the matrix, there was a chance that for the 'L' shape, the final column which contains the tail of the 'L' would be cut off, producing an misclassified dataset. By ensuring that the shift value cannot produce invalid datasets as well as having the option to discretely check if the dataset is what we expect, we avoid this problem.