Add generate_minimal_complete_dataset(), dataset shuffling
Added ability to generate a minimal dataset containing every possible combination of the 'I' and 'L' matrices. Similarly, added dataset shuffling which should be done on the dataset regardless as stated in today's lecture. Fixed a exception which would be caused in the number of folds was greater than the number of elements in the dataset.