Add generate_dataset(), set epoch's to 1
Added a dataset generator which supports generating a dataset which contains a specific amount of 'I' and 'L' matrices. Also set epoch's to 1 (the amount of times the Perceptron is trained on the dataset, was 500 before), as Perceptron should not be run multiple times over the same dataset to match the expectations for the lab.

The program is now more or less setup for the requirements for the lab.