Implement arbitrary down-right shifting, remove optimal finder, clarify matrix-dataset verbiage
Implemented the ability to shift the character in the matrix to the right and downwards. Removed all the code related to finding an optimal dataset size for an accuracy as it was cluttering the codebase and not required (plus not working with the new shifting system). Made it clear when we are working with datasets (lists of matrices) and singular matrices.